requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
boost=1.74.0=py310h7c3ba0c_5
boost-cpp=1.74.0=h6cacc03_7
brotli-python=1.0.9=py310hd8f1fbe_9
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cairo=1.16.0=ha00ac49_1009
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
coverage=7.3.0=py310h2372a71_0
cycler=0.11.0=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
filelock=3.12.2=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=hef1e5e3_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=hca18f0e_2
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
gmp=6.2.1=h58526e2_0
gmpy2=2.1.2=py310h3ec546c_1
googledrivedownloader=0.4=pyhd3deb0d_1
greenlet=2.0.2=py310hc6cd4ac_1
icu=69.1=h9c3ff4c_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
iniconfig=2.0.0=pyhd8ed1ab_0
isodate=0.6.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
kiwisolver=1.4.4=py310hbf28c38_1
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=17_linux64_openblas
libcblas=3.9.0=17_linux64_openblas
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.4=hebfc3b9_0
libgomp=13.1.0=he5830b7_0
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=17_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.23=pthreads_h80387f5_0
libpng=1.6.39=h753d276_0
libprotobuf=3.21.12=h3eb15da_0
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=hd590300_1
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h0b41bf4_0
libxml2=2.9.14=haae042b_4
libzlib=1.2.13=hd590300_5
markupsafe=2.1.3=py310h2372a71_0
matplotlib-base=3.4.3=py310h23f4a51_2
mendeleev=0.14.0=pyha21a80b_0
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.0=hb012696_0
mpmath=1.3.0=pyhd8ed1ab_0
ncurses=6.4=hcb278e6_0
networkx=3.1=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=1.25.2=py310ha4c1d20_0
omegaconf=2.3.0=pyhd8ed1ab_0
openssl=3.1.2=hd590300_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py310h7cbd5c2_1
pcre2=10.40=hc3806b6_0
pillow=9.4.0=py310h6a678d5_0
pip=23.2.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
platformdirs=3.10.0=pyhd8ed1ab_0
pluggy=1.2.0=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
pthread-stubs=0.4=h36c2ea0_1001
pycairo=1.24.0=py310hda9f760_0
pyfiglet=0.8.post1=py_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pytest=7.4.0=pyhd8ed1ab_0
pytest-cov=4.1.0=pyhd8ed1ab_0
python=3.10.12=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python-wget=3.2=py_0
python_abi=3.10=3_cp310
pytorch=2.0.0=cpu_generic_py310h7ffd2bf_1
pytorch_geometric=2.1.0.post1=pyhd8ed1ab_0
pytorch_scatter=2.1.1=cpu_py310hf5a12b5_1
pytorch_sparse=0.6.17=py310he48728a_1
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_5
rdflib=7.0.0=pyhd8ed1ab_0
rdkit=2022.03.5=py310h1c297d8_0
readline=8.2=h8228510_1
reportlab=3.6.12=py310h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.3.0=py310hf7d194e_0
scipy=1.11.2=py310ha4c1d20_0
setuptools=68.1.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
sqlalchemy=2.0.20=py310h2372a71_0
sympy=1.12=pypyh9d50eac_103
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.3.3=py310h2372a71_0
tqdm=4.66.1=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.41.1=pyhd8ed1ab_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.6=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.2=hfc55251_7
